initial split between lib and ui. It type checks, but I had to remove a lot of UI code, which I have to add back in now
