+++ troubadour_lib/.gitignore
@@ -1,0 +1,11 @@
+example.log
+.vscode
+__pycache__
+py_venv
+
+# Added by cargo
+get_table_top_sounds
+
+/target
+/saves
+troubadour
